Move and rename Packet types
